Rename `Response` to `Outcome` (#65)
Naming is hard and I've slowly come to the conclusion that using
`Response` for state handlers output wasn't the best choice. The term
"response" is already very common within programming and often implies
an associated "request", which isn't the case in statig. I've also found
myself in situations where I wanted to store (application-level)
requests within local-storage, which resulted in function signatures
like this:

```rust
#[state]
fn processing(&mut self, request: Request, event: &Event) -> Response<State> { ... }
```

Readers of this signature would be forgiven for thinking `Request` and
`Response` are somehow associated with each other, even though this
isn't the case.

After some consideration (and consulting of thesaurus.com) I've landed
on `Outcome`. I think it is a more natural fit to the state machine
terminology (an "event" produces an "outcome") and isn't something that
would cause other conflicts (i.e. something like `Result`).

This is a breaking change, but to make the transition as pain-free as
possible I've provided a `Response` type alias that is marked as
deprecated and can be removed in some future release.